+++ HISTORY.rst
@@ -10,6 +10,7 @@ develop
 - Group folders on top in files browser
 - Sort folder and file names case insensitive
 - Reintroduce headers in settings page of new schematisation wizard
+- Run auto refresh on regaining window focus
 - Make breadcrumbs widget clearer with dropdown for long paths